pyproject.toml
[tool.poetry]
name = "bartender-bot"
version = "0.1.0"
description = ""
authors = ["Daniel Malik <mail@fronbasal.de>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.1"
requests = "^2.25.1"
dynaconf = "^3.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
